Refactor: Update PDF generation and data models for sales
This commit introduces several changes related to PDF generation and data models for sales, including:
- Updated `SaleDataEntity` to change the `number` field type from `Int` to `Long`.
- Updated `DatabaseModule` to enable destructive migration for the Room database.
- Updated `EmissionResponseDto` to change the `number` field type from `Int` to `Long`.
- Incremented database version from 2 to 3 in `DataBaseApp.kt`.
- In `PdfManagerRepository`:
    - Modified `generatePreview` to accept an `isLetter` boolean and return `ResultType<DocumentElectronic>`. It now generates either a letter or 80mm format PDF.
    - Renamed `generateDte` to `generateDteLetter` and updated its `number` parameter to `Long`.
    - Added a new method `generateDte80mm` to create PDFs in 80mm format. This method includes logic for layout, content, and optional fiscal stamp.
- In `OpenFacturaRepository`:
    - Updated `generateInvoice` to accept an `isLetter` boolean. It now calls either `generateDteLetter` or `generateDte80mm` based on this flag.
    - The `DocumentElectronic` returned by `generateInvoice` now includes `docType`, `format`, `token`, `resolution`, `timbre`, `sale`, and `payment`.
    - Ensured `saveReportSAle` correctly maps `response.number` to `SaleDataEntity.number`.
- Commented out the `printTooth` dependency in `gradle/libs.versions.toml`.

+++ app/src/main/java/com/blipblipcode/distribuidoraayl/core/local/entities/openFactura/reportSale/SaleDataEntity.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class SaleDataEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0L,
-    val number: Int,
+    val number: Long,
     val clientRut: String,
     val date: String,
     val token: String?,
